cleanup react code.
