+++ src/react/interfaces/CRMEvent.ts
@@ -1,0 +1,11 @@
+interface CRMEvent {
+    Desc: string,
+    End: string,
+    Id: number,
+    InActive: boolean,
+    Start: string,
+    Text: string,
+    Type: number,
+    Title: string
+  }
+export default CRMEvent
