+++ src/react/window-context-service.jsx
@@ -1,0 +1,1 @@
+export default window.CRM.root;
